Move Reciter and Recitation models to recite app
