+++ config/settings/base.py
@@ -31,6 +31,7 @@ DJANGO_APPS = [
 THIRD_PARTY_APPS = []
 LOCAL_APPS = [
     'quran',
+    'recite',
 ]
 # Application definition
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
